Master

+++ backend/controllers/AktMitraBisnisAlamatController.php
@@ -9,6 +9,7 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use backend\models\AktKota;
+
 /**
  * AktMitraBisnisAlamatController implements the CRUD actions for AktMitraBisnisAlamat model.
  */
@@ -98,7 +99,7 @@ class AktMitraBisnisAlamatController extends Controller
         $total_kota = AktKota::find()->count();
         $nomor_kota = "KT" . str_pad($total_kota + 1, 3, "0", STR_PAD_LEFT);
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['akt-mitra-bisnis/view', 'id' => $model->id_mitra_bisnis,'#' => 'alamat']);
+            return $this->redirect(['akt-mitra-bisnis/view', 'id' => $model->id_mitra_bisnis, '#' => 'alamat']);
         }
 
         return $this->render('update', [
@@ -118,7 +119,41 @@ class AktMitraBisnisAlamatController extends Controller
     public function actionDelete($id)
     {
         $model = $this->findModel($id);
-        $model->delete();
+
+        $db = Yii::$app->getDb();
+        $dbname = 'dbname';
+        $dsn = $db->dsn;
+        $dbName = getDsnAttribute($dbname, $dsn);
+        $columnName = 'id_mitra_bisnis_alamat';
+        $thisTable = 'akt_mitra_bisnis_alamat';
+        $thisId = $model->id_mitra_bisnis_alamat;
+        $rows = (new \yii\db\Query())
+            ->select(['TABLE_NAME'])
+            ->from('INFORMATION_SCHEMA.COLUMNS')
+            ->where(['TABLE_SCHEMA' => $dbName])
+            ->andWhere(['COLUMN_NAME' => $columnName])
+            ->andWhere(['!=', 'TABLE_NAME', $thisTable])
+            ->all();
+        $array_table_name = array();
+        $totalan_countData = 0;
+        foreach ($rows as $key => $value) {
+            # code...
+            $rows2 = (new \yii\db\Query())
+                ->select(['COUNT(*) as countData'])
+                ->from($value['TABLE_NAME'])
+                ->where([$columnName => $thisId])
+                ->one();
+            $array_table_name[] = $value['TABLE_NAME'] . ' - ' . $rows2['countData'];
+            $totalan_countData += $rows2['countData'];
+        }
+
+        if ($totalan_countData == 0) {
+            # code...
+            $model->delete();
+        } else {
+            # code...
+            Yii::$app->session->setFlash('warning', [['Perhatian!', '<i class="fa fa-exclamation-triangle" aria-hidden="true"></i> Data Alamat : <b>' . $model->keterangan_alamat . '</b> Tidak Dapat Di Hapus, Dikarenakan Data Masih Digunakan!']]);
+        }
 
         return $this->redirect(['akt-mitra-bisnis/view', 'id' => $model->id_mitra_bisnis, '#' => 'alamat']);
     }
