+++ backend/controllers/AktMitraBisnisKontakController.php
@@ -68,7 +68,7 @@ class AktMitraBisnisKontakController extends Controller
         $model->id_mitra_bisnis = $_GET['id'];
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['akt-mitra-bisnis/view', 'id' => $model->id_mitra_bisnis,'#' => 'kontak']);
+            return $this->redirect(['akt-mitra-bisnis/view', 'id' => $model->id_mitra_bisnis, '#' => 'kontak']);
         }
 
         return $this->render('create', [
@@ -88,7 +88,7 @@ class AktMitraBisnisKontakController extends Controller
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['akt-mitra-bisnis/view', 'id' => $model->id_mitra_bisnis,'#' => 'kontak']);
+            return $this->redirect(['akt-mitra-bisnis/view', 'id' => $model->id_mitra_bisnis, '#' => 'kontak']);
         }
 
         return $this->render('update', [
@@ -106,9 +106,43 @@ class AktMitraBisnisKontakController extends Controller
     public function actionDelete($id)
     {
         $model = $this->findModel($id);
-        $model->delete();
 
-        return $this->redirect(['akt-mitra-bisnis/view', 'id' => $model->id_mitra_bisnis,'#' => 'kontak']);
+        $db = Yii::$app->getDb();
+        $dbname = 'dbname';
+        $dsn = $db->dsn;
+        $dbName = getDsnAttribute($dbname, $dsn);
+        $columnName = 'id_mitra_bisnis_kontak';
+        $thisTable = 'akt_mitra_bisnis_kontak';
+        $thisId = $model->id_mitra_bisnis_kontak;
+        $rows = (new \yii\db\Query())
+            ->select(['TABLE_NAME'])
+            ->from('INFORMATION_SCHEMA.COLUMNS')
+            ->where(['TABLE_SCHEMA' => $dbName])
+            ->andWhere(['COLUMN_NAME' => $columnName])
+            ->andWhere(['!=', 'TABLE_NAME', $thisTable])
+            ->all();
+        $array_table_name = array();
+        $totalan_countData = 0;
+        foreach ($rows as $key => $value) {
+            # code...
+            $rows2 = (new \yii\db\Query())
+                ->select(['COUNT(*) as countData'])
+                ->from($value['TABLE_NAME'])
+                ->where([$columnName => $thisId])
+                ->one();
+            $array_table_name[] = $value['TABLE_NAME'] . ' - ' . $rows2['countData'];
+            $totalan_countData += $rows2['countData'];
+        }
+
+        if ($totalan_countData == 0) {
+            # code...
+            $model->delete();
+        } else {
+            # code...
+            Yii::$app->session->setFlash('warning', [['Perhatian!', '<i class="fa fa-exclamation-triangle" aria-hidden="true"></i> Data Kontak : <b>' . $model->nama_kontak . '</b> Tidak Dapat Di Hapus, Dikarenakan Data Masih Digunakan!']]);
+        }
+
+        return $this->redirect(['akt-mitra-bisnis/view', 'id' => $model->id_mitra_bisnis, '#' => 'kontak']);
     }
 
     /**
